requirements.txt
asgiref==3.9.1
Django==5.2.5
django-admin-interface==0.30.1
django-colorfield==0.14.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
